move map_array to util
