+++ skimage/util/_remap.pyx
